Complted MVC Structure,app running fine
